Update SimpleServer.java
Since it's supposed to be public, why not bind on all interfaces?  I'm having a problem connecting to a server that is getting an ipv6 address, but my tools can't reach it.  Since there is no configuration to this on the server side, just bind to all interfaces and let the user figure out which ips the device has.